+++ frontend/src/App.tsx
@@ -2,8 +2,9 @@ import ProtectedRoute from "@/components/ProtectedRoute";
 import DashboardRoute from "@/routes/Dashboard";
 import DashboardLayout from "@/routes/DashboardLayout";
 import LoginRoute from "@/routes/Login";
+import ProjectRoute from "@/routes/Project";
 import RegisterRoute from "@/routes/Register";
-import { BrowserRouter, Route, Routes } from "react-router";
+import { BrowserRouter, Navigate, Route, Routes } from "react-router";
 
 function App() {
   return (
@@ -18,10 +19,11 @@ function App() {
           }
         >
           <Route index element={<DashboardRoute />} />
-          <Route path="project/:id" element={<div>Project Route</div>} />
+          <Route path="project/:id" element={<ProjectRoute />} />
         </Route>
         <Route path="/login" element={<LoginRoute />} />
         <Route path="/register" element={<RegisterRoute />} />
+        <Route path="*" element={<Navigate to="/dashboard" replace />} />
       </Routes>
     </BrowserRouter>
   );